+++ assets/javascripts/kitten/components/buttons/button/stories.js
@@ -69,7 +69,7 @@ const info = {
     ~~~js
     <Button tiny>MyButton</Button>
     <Button big>MyButton</Button>
-    <Button giant>MyButton</Button>
+    <Button huge>MyButton</Button>
     ~~~
 
     #### Display
@@ -141,7 +141,7 @@ storiesOf('Buttons/Button', module)
                 <Button
                   tiny={boolean('Tiny', false)}
                   big={boolean('Big', false)}
-                  giant={boolean('Giant', false)}
+                  huge={boolean('Huge', false)}
                   fluid={boolean('Fluid', false)}
                   modifier={select('Modifier', modifierOptions, 'hydrogen')}
                   disabled={boolean('Disabled', false)}
@@ -170,7 +170,7 @@ storiesOf('Buttons/Button', module)
                 <Button
                   tiny={boolean('Tiny', false)}
                   big={boolean('Big', false)}
-                  giant={boolean('Giant', false)}
+                  huge={boolean('Huge', false)}
                   fluid={boolean('Fluid', false)}
                   modifier={select('Modifier', modifierOptions, 'hydrogen')}
                   disabled={boolean('Disabled', false)}
@@ -211,7 +211,7 @@ storiesOf('Buttons/Button', module)
                 <Button
                   tiny={boolean('Tiny', false)}
                   big={boolean('Big', false)}
-                  giant={boolean('Giant', false)}
+                  huge={boolean('Huge', false)}
                   modifier={select('Modifier', modifierOptions, 'hydrogen')}
                   fluid={boolean('Fluid', false)}
                   disabled={boolean('Disabled', false)}
@@ -240,7 +240,7 @@ storiesOf('Buttons/Button', module)
               <Button
                 tiny={boolean('Tiny', false)}
                 big={boolean('Big', false)}
-                giant={boolean('Giant', false)}
+                huge={boolean('Huge', false)}
                 modifier={select('Modifier', modifierOptions, 'helium')}
                 disabled={boolean('Disabled', false)}
               >